backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.1.16
celery==5.3.6
redis==5.0.1
boto3==1.34.34
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=2.10.0
groq>=0.4.2
httpx>=0.27.2
requests>=2.31.0

